Check online status before prefetching URLs
Related to issue #73

As a website owner, I want to only prefetch URLs when I am online, so that I am not getting network errors when my site is offline or served by a service-worker.

Changes:

- Implement isOnline function to check online status via navigator.onLine (https://developer.mozilla.org/de/docs/Web/API/NavigatorOnLine/onLine).
- Implement guard in preload function to check online status with this function.

Only tested on my setup. Has to be tested on larger scope before merging.