Prefetch with high fetchPriority on mouse & touch events
By default, a prefetch is loaded with a low priority.
When there’s a fair chance that this prefetch is going to be used in the near term (= after a touch/mouse event), giving it a high priority would help make the page load faster in case there are other resources loading.

Prioritizing it implicitly means deprioritizing every other resource that’s loading on the page.
Due to HTML documents usually being much smaller than other resources (notably images and JavaScript), this theft of bandwidth should rarely be detrimental.

Closes #117