Make module importable; data-instant-config
It now needs to have a data-instant-config attribute on the script element to init, OR it can be import()ed then init()ed.

This will allow to have a more flexible configuration system later on, including one that can be updated after initial page load.

This prevents the (unsupported and problematic) possibility of loading it as a classic script, which some webmasters do.